microservices/document_service/app/requirements.txt
# Document Service Requirements
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiohttp>=3.8.1
pymongo>=4.0.1
motor>=2.5.1
python-dotenv>=0.19.1
minio>=7.1.0
kafka-python>=2.0.2
pydantic>=1.9.0
aiofiles>=0.8.0
httpx>=0.22.0
langchain>=0.0.142
# PDF processing dependencies
PyPDF2>=2.0.0  # Added PyPDF2 for PDF text extraction
pdf2image>=1.16.0
pytesseract>=0.3.9
Pillow>=9.0.0  # Added explicit Pillow dependency for image processing
numpy>=1.20.0

# OpenTelemetry dependencies - using stable versions without beta requirements
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp>=1.20.0
# Removing problematic aiohttp instrumentation
# opentelemetry-instrumentation-aiohttp
opentelemetry-instrumentation-requests
python-json-logger>=2.0.7
elasticsearch[async]>=8.0.0

# Dependencies added by fix script
nltk>=3.8.1
scikit-learn>=1.3.0
spacy>=3.6.1
